requirements.txt
streamlit 
torch 
pillow 
transformers 
scikit-learn 
numpy